tabl

+++ src/main/java/hello/DoctorRepository.java
@@ -5,5 +5,5 @@ import hello.model.Doctor;
 /**
  * Created by elvir_000 on 17.05.2017.
  */
-public interface DoctorRepository extends CrudRepository<Doctor, Long> {
+public interface DoctorRepository extends CrudRepository<Doctor, Integer> {
 }
